fix budget type in dtos

+++ src/main/java/com/codeseek/footballmanager/dto/team/TeamResponseDto.java
@@ -1,10 +1,11 @@
 package com.codeseek.footballmanager.dto.team;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record TeamResponseDto(
     UUID id,
     String name,
-    double budget,
+    BigDecimal budget,
     double commissionRate
 ) {}
